backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.6.0
drf-spectacular>=0.26.0
django-filter>=23.0.0
Pillow>=9.5.0
celery>=5.2.0
redis>=4.5.0
django-extensions>=3.2.0
python-dateutil>=2.8.0
openpyxl>=3.1.0
pandas>=2.0.0
